requirements.txt
PyCharacterAI
requests
python-dotenv
pygame